Use size_t for sizes in CFX_BitStream.
As would be flagged by -Wshorten-64-to-32

-- beef up some tests.

Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/86892
Commit-Queue: Tom Sepez <tsepez@chromium.org>

+++ core/fxcrt/cfx_bitstream.h
@@ -7,6 +7,7 @@
 #ifndef CORE_FXCRT_CFX_BITSTREAM_H_
 #define CORE_FXCRT_CFX_BITSTREAM_H_
 
+#include <stddef.h>
 #include <stdint.h>
 
 #include "core/fxcrt/unowned_ptr.h"
@@ -20,19 +21,19 @@ class CFX_BitStream {
   void ByteAlign();
 
   bool IsEOF() const { return m_BitPos >= m_BitSize; }
-  uint32_t GetPos() const { return m_BitPos; }
+  size_t GetPos() const { return m_BitPos; }
   uint32_t GetBits(uint32_t nBits);
 
-  void SkipBits(uint32_t nBits) { m_BitPos += nBits; }
+  void SkipBits(size_t nBits) { m_BitPos += nBits; }
   void Rewind() { m_BitPos = 0; }
 
-  uint32_t BitsRemaining() const {
+  size_t BitsRemaining() const {
     return m_BitSize >= m_BitPos ? m_BitSize - m_BitPos : 0;
   }
 
  private:
-  uint32_t m_BitPos = 0;
-  const uint32_t m_BitSize;
+  size_t m_BitPos = 0;
+  const size_t m_BitSize;
   UnownedPtr<const uint8_t> const m_pData;
 };
